backend/pyproject.toml
[tool.poetry]
name = "ledger-sync"
version = "0.1.0"
description = "Production-ready data ingestion and reconciliation engine for Money Manager Pro Excel exports"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "ledger_sync", from = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
sqlalchemy = "^2.0.0"
alembic = "^1.13.0"
pandas = "^2.2.0"
openpyxl = "^3.1.0"
typer = "^0.12.0"
pydantic = "^2.7.0"
pydantic-settings = "^2.2.0"
rich = "^13.7.0"
fastapi = "^0.115.0"
uvicorn = {extras = ["standard"], version = "^0.32.0"}
python-multipart = "^0.0.9"
psycopg2-binary = "^2.9.9"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
bcrypt = "^4.0.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.1.0"
pytest-cov = "^5.0.0"
ruff = "^0.4.0"
black = "^24.4.0"
mypy = "^1.10.0"
pandas-stubs = "^2.2.0"
types-openpyxl = "^3.1.0"

[tool.poetry.scripts]
ledger-sync = "ledger_sync.cli.main:app"

[tool.black]
line-length = 100
target-version = ['py312']

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.ruff.lint.per-file-ignores]
# Allow asserts in tests
"tests/**/*.py" = ["S101"]
# Migration files are auto-generated
"**/migrations/**/*.py" = ["D100", "D103", "D104", "D400", "D415"]
# FastAPI dependency injection uses Depends() in defaults - this is the standard pattern
# Boolean params in Query() are also standard FastAPI patterns
# PERF401 - conditional appends in loops are more readable than list.extend alternatives
"src/ledger_sync/api/**/*.py" = ["B008", "FBT001", "FBT002", "FBT003", "PERF401"]
# CLI uses typer.Argument/Option in defaults - this is the standard pattern
"src/ledger_sync/cli/**/*.py" = ["B008", "FBT001", "FBT002", "FBT003"]
# Core module has boolean parameter in sync_engine.import_file which is intentional
# PERF401 - conditional appends in loops are more readable
"src/ledger_sync/core/sync_engine.py" = ["FBT001", "FBT002"]
"src/ledger_sync/core/analytics_engine.py" = ["PERF401"]

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
plugins = ["pydantic.mypy"]

[[tool.mypy.overrides]]
module = "openpyxl.*"
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
